requirements.txt
fastapi
uvicorn
sqlalchemy
Jinja2
alembic
python-jose 
cryptography
passlib 
bcrypt 
python-multipart